Moved to a simpler workflow parsing mechanism
