+++ app/lib/bk/compat/parsers/bitrise.rb
@@ -4,7 +4,6 @@ require_relative '../translator'
 require_relative '../models/pipeline'
 require_relative 'bitrise/steps'
 require_relative 'bitrise/workflows'
-require_relative 'bitrise/loader'
 
 module BK
   module Compat
@@ -32,14 +31,11 @@ module BK
       def initialize(text, options = {})
         @config = YAML.safe_load(text, aliases: true)
         @options = options
-        @steps_by_key = {}
 
         register_translators!
       end
 
       def parse
-        load_workflows!
-
         workflows = @config.fetch('workflows', {})
         bk_steps = workflows.map { |wf_name, wf_config| parse_workflow(wf_name, wf_config) }
 
@@ -53,10 +49,6 @@ module BK
       def register_translators!
         @translator = BK::Compat::StepTranslator.new
         @translator.register(BK::Compat::BitriseSteps::Translator.new)
-      end
-
-      def load_workflows!
-        @config.fetch('workflows', {}).map { |key, config| load_workflow(key, config) }
       end
     end
   end
